Version 1.0

+++ TaxiService/TaxiService/Models/Location.cs
@@ -43,9 +43,9 @@ namespace TaxiService.Models
 
         public int Id { get; set; }
 
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
 
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
         public string Street { get; set; }
